rename sdxl_support.pt, move SDXLExportWrapper
